+++ src/container/grid.rs
@@ -1,6 +1,7 @@
 use egui::{emath::Rangef, pos2, vec2, NumExt as _, Rect};
 use itertools::Itertools as _;
 
+use crate::behavior::EditAction;
 use crate::{
     Behavior, ContainerInsertion, DropContext, InsertionPoint, ResizeState, SimplifyAction, TileId,
     Tiles, Tree,
@@ -160,7 +161,10 @@ impl Grid {
             (num_cols, num_rows)
         };
 
-        debug_assert!(visible_children_and_holes.len() <= num_cols * num_rows);
+        debug_assert!(
+            visible_children_and_holes.len() <= num_cols * num_rows,
+            "Bug in egui_tiles::Grid::layout"
+        );
 
         // Figure out where each column and row goes:
         self.col_shares.resize(num_cols, 1.0);
@@ -169,8 +173,16 @@ impl Grid {
         let col_widths = sizes_from_shares(&self.col_shares, rect.width(), gap);
         let row_heights = sizes_from_shares(&self.row_shares, rect.height(), gap);
 
-        debug_assert_eq!(col_widths.len(), num_cols);
-        debug_assert_eq!(row_heights.len(), num_rows);
+        debug_assert_eq!(
+            col_widths.len(),
+            num_cols,
+            "Bug in egui_tiles::Grid::layout"
+        );
+        debug_assert_eq!(
+            row_heights.len(),
+            num_rows,
+            "Bug in egui_tiles::Grid::layout"
+        );
 
         {
             let mut x = rect.left();
@@ -189,8 +201,16 @@ impl Grid {
             }
         }
 
-        debug_assert_eq!(self.col_ranges.len(), num_cols);
-        debug_assert_eq!(self.row_ranges.len(), num_rows);
+        debug_assert_eq!(
+            self.col_ranges.len(),
+            num_cols,
+            "Bug in egui_tiles::Grid::layout"
+        );
+        debug_assert_eq!(
+            self.row_ranges.len(),
+            num_rows,
+            "Bug in egui_tiles::Grid::layout"
+        );
 
         // Layout each child:
         for (i, &child) in visible_children_and_holes.iter().enumerate() {
@@ -223,7 +243,7 @@ impl Grid {
         tree: &mut Tree<Pane>,
         behavior: &mut dyn Behavior<Pane>,
         drop_context: &mut DropContext,
-        ui: &mut egui::Ui,
+        ui: &egui::Ui,
         tile_id: TileId,
     ) {
         for &child in &self.children {
@@ -246,33 +266,35 @@ impl Grid {
             );
         }
 
-        self.resize_columns(&mut tree.tiles, behavior, ui, tile_id);
-        self.resize_rows(&mut tree.tiles, behavior, ui, tile_id);
+        self.resize_columns(&tree.tiles, behavior, ui, tile_id);
+        self.resize_rows(&tree.tiles, behavior, ui, tile_id);
     }
 
     fn resize_columns<Pane>(
         &mut self,
-        tiles: &mut Tiles<Pane>,
+        tiles: &Tiles<Pane>,
         behavior: &mut dyn Behavior<Pane>,
-        ui: &mut egui::Ui,
+        ui: &egui::Ui,
         parent_id: TileId,
     ) {
-        let parent_rect = tiles.rect(parent_id);
+        let parent_rect = tiles.rect_or_die(parent_id);
         for (i, (left, right)) in self.col_ranges.iter().copied().tuple_windows().enumerate() {
-            let resize_id = egui::Id::new((parent_id, "resize_col", i));
+            let resize_id = ui.id().with((parent_id, "resize_col", i));
 
             let x = egui::lerp(left.max..=right.min, 0.5);
 
             let mut resize_state = ResizeState::Idle;
-            if let Some(pointer) = ui.ctx().pointer_latest_pos() {
-                let line_rect = Rect::from_center_size(
-                    pos2(x, parent_rect.center().y),
-                    vec2(
-                        2.0 * ui.style().interaction.resize_grab_radius_side,
-                        parent_rect.height(),
-                    ),
-                );
-                let response = ui.interact(line_rect, resize_id, egui::Sense::click_and_drag());
+            let line_rect = Rect::from_center_size(
+                pos2(x, parent_rect.center().y),
+                vec2(
+                    2.0 * ui.style().interaction.resize_grab_radius_side,
+                    parent_rect.height(),
+                ),
+            );
+            let response = ui.interact(line_rect, resize_id, egui::Sense::click_and_drag());
+            // NOTE: Check for interaction with line_rect BEFORE entering the 'IF block' below,
+            // otherwise we miss the start of a drag event in certain cases (e.g. touchscreens).
+            if let Some(pointer) = ui.ctx().pointer_interact_pos() {
                 resize_state = resize_interaction(
                     behavior,
                     &self.col_ranges,
@@ -294,27 +316,29 @@ impl Grid {
 
     fn resize_rows<Pane>(
         &mut self,
-        tiles: &mut Tiles<Pane>,
+        tiles: &Tiles<Pane>,
         behavior: &mut dyn Behavior<Pane>,
-        ui: &mut egui::Ui,
+        ui: &egui::Ui,
         parent_id: TileId,
     ) {
-        let parent_rect = tiles.rect(parent_id);
+        let parent_rect = tiles.rect_or_die(parent_id);
         for (i, (top, bottom)) in self.row_ranges.iter().copied().tuple_windows().enumerate() {
-            let resize_id = egui::Id::new((parent_id, "resize_row", i));
+            let resize_id = ui.id().with((parent_id, "resize_row", i));
 
             let y = egui::lerp(top.max..=bottom.min, 0.5);
 
             let mut resize_state = ResizeState::Idle;
-            if let Some(pointer) = ui.ctx().pointer_latest_pos() {
-                let line_rect = Rect::from_center_size(
-                    pos2(parent_rect.center().x, y),
-                    vec2(
-                        parent_rect.width(),
-                        2.0 * ui.style().interaction.resize_grab_radius_side,
-                    ),
-                );
-                let response = ui.interact(line_rect, resize_id, egui::Sense::click_and_drag());
+            let line_rect = Rect::from_center_size(
+                pos2(parent_rect.center().x, y),
+                vec2(
+                    parent_rect.width(),
+                    2.0 * ui.style().interaction.resize_grab_radius_side,
+                ),
+            );
+            let response = ui.interact(line_rect, resize_id, egui::Sense::click_and_drag());
+            // NOTE: Check for interaction with line_rect BEFORE entering the 'IF block' below,
+            // otherwise we miss the start of a drag event in certain cases (e.g. touchscreens).
+            if let Some(pointer) = ui.ctx().pointer_interact_pos() {
                 resize_state = resize_interaction(
                     behavior,
                     &self.row_ranges,
@@ -379,7 +403,7 @@ fn resize_interaction<Pane>(
     dx: f32,
     i: usize,
 ) -> ResizeState {
-    assert_eq!(ranges.len(), shares.len());
+    assert_eq!(ranges.len(), shares.len(), "Bug in egui_tiles::Grid");
     let num = ranges.len();
     let tile_width = |i: usize| ranges[i].span();
 
@@ -387,7 +411,7 @@ fn resize_interaction<Pane>(
     let right = i + 1;
 
     if splitter_response.double_clicked() {
-        behavior.on_edit();
+        behavior.on_edit(EditAction::TileResized);
 
         // double-click to center the split between left and right:
         let mean = 0.5 * (shares[left] + shares[right]);
@@ -395,7 +419,7 @@ fn resize_interaction<Pane>(
         shares[right] = mean;
         ResizeState::Hovering
     } else if splitter_response.dragged() {
-        behavior.on_edit();
+        behavior.on_edit(EditAction::TileResized);
 
         if dx < 0.0 {
             // Expand right, shrink stuff to the left:
@@ -507,6 +531,14 @@ mod tests {
             }
 
             fn tab_title_for_pane(&mut self, _pane: &Pane) -> egui::WidgetText {
+                panic!()
+            }
+
+            fn is_tab_closable(&self, _tiles: &Tiles<Pane>, _tile_id: TileId) -> bool {
+                panic!()
+            }
+
+            fn on_tab_close(&mut self, _tiles: &mut Tiles<Pane>, _tile_id: TileId) -> bool {
                 panic!()
             }
         }